# swagger.yaml
openapi: 3.0.0
info:
  title: Fitness API
  description: |
    API RESTful complète pour l'application mobile Fitness.
    Supporte gestion des utilisateurs, courses, séances, pas, poids, amis, messages et badges.
  version: 1.0.0
  contact:
    name: API Support
    email: support@fitnesssapi.com
  license:
    name: MIT
servers:
  - url: https://fitnesssapi.onrender.com
    description: Production Server
  - url: http://localhost:3000
    description: Local Development Server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Authentification par Bearer Token (JWT)
  schemas:
...
    RunningSession:
      type: object
      properties:
        id:
          type: string
          format: uuid
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        durationSeconds:
          type: integer
          example: 3600
        distanceKm:
          type: number
          example: 10.5
        avgSpeedKmh:
          type: number
          example: 10.5
        maxSpeedKmh:
          type: number
          example: 12
        caloriesBurned:
          type: number
          example: 750
        elevationGainM:
          type: number
          example: 150
        avgHeartRate:
          type: integer
          example: 140
        maxHeartRate:
          type: integer
          example: 165
        weather:
          type: string
          nullable: true
          example: sunny
        temperatureC:
          type: number
          nullable: true
          example: 22
        notes:
          type: string
          nullable: true
        isCompleted:
          type: boolean
        userId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        gpsPoints:
          type: array
          items:
            $ref: '#/components/schemas/GPSPoint'
...
    GPSPoint:
      type: object
      properties:
        id:
          type: string
          format: uuid
        latitude:
          type: number
          example: 48.8566
        longitude:
          type: number
          example: 2.3522
        altitudeM:
          type: number
          nullable: true
        speedKmh:
          type: number
          nullable: true
        accuracyM:
          type: number
          nullable: true
        recordedAt:
          type: string
          format: date-time
...
    Workout:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: Upper Body
        workoutType:
          type: string
          example: strength
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        durationMinutes:
          type: integer
        caloriesBurned:
          type: number
        notes:
          type: string
          nullable: true
        exerciseCount:
          type: integer
        userId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        exerciseLogs:
          type: array
          items:
            $ref: '#/components/schemas/ExerciseLog'
...
    WorkoutUpdate:
      type: object
      properties:
        name:
          type: string
        workoutType:
          type: string
        durationMinutes:
          type: integer
        caloriesBurned:
          type: number
        notes:
          type: string
    ExerciseLog:
      type: object
      properties:
        id:
          type: string
          format: uuid
        exerciseId:
          type: string
        exerciseName:
          type: string
        exerciseType:
          type: string
        notes:
          type: string
          nullable: true
        setLogs:
          type: array
          items:
            $ref: '#/components/schemas/SetLog'
        cardioLogs:
          type: array
          items:
            $ref: '#/components/schemas/CardioLog'
...
    SetLog:
      type: object
      properties:
        id:
          type: string
          format: uuid
        setNumber:
          type: integer
        reps:
          type: integer
        weightKg:
          type: number
        isCompleted:
          type: boolean
        restTimeSec:
          type: integer
...
    CardioLog:
      type: object
      properties:
        id:
          type: string
          format: uuid
        durationMinutes:
          type: number
        distanceKm:
          type: number
        avgSpeedKmh:
          type: number
        maxSpeedKmh:
          type: number
        caloriesBurned:
          type: number
        avgHeartRate:
          type: integer
        maxHeartRate:
          type: integer
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Code d'erreur
          example: VALIDATION_ERROR
        message:
          type: string
          description: Message d'erreur lisible
          example: Email or password incorrect
...
paths:
...
  /running/{id}:
    get:
      tags:
        - Running
      summary: Récupérer une session
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Détails de la session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunningSession'
        '404':
          description: Session non trouvée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - Running
      summary: Supprimer une session
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Session supprimée
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string

  # ===== WORKOUTS =====
...
  /workouts/{id}:
    get:
      tags:
        - Workouts
      summary: Récupérer une séance par ID
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Séance trouvée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workout'
        '404':
          description: Séance non trouvée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    patch:
      tags:
        - Workouts
      summary: Mettre à jour une séance
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkoutUpdate'
      responses:
        '200':
          description: Séance mise à jour
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workout'
        '404':
          description: Séance non trouvée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - Workouts
      summary: Supprimer une séance
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Séance supprimée
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '404':
          description: Séance non trouvée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ===== CUSTOM PROGRAMS =====
...